requirements.txt
astroid==2.4.2
future==0.18.2
isort==4.3.21
lazy-object-proxy==1.4.3
llvmlite==0.34.0
mccabe==0.6.1
numba==0.51.1
numpy==1.19.1
opencv-python==4.4.0.42
Pillow==7.2.0
ply==3.11
pybullet==2.8.4
pylint==2.5.3
scipy==1.5.2
six==1.15.0
toml==0.10.1
torch==1.6.0
torchvision==0.7.0
tqdm==4.51.0
transforms3d==0.3.1
trimesh==3.8.4
typed-ast==1.4.1
wrapt==1.12.1
